Creating alarm view
