+++ DeFuge/AppDelegate.swift
@@ -16,6 +16,19 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
         // Override point for customization after application launch.
+        
+        // TODO: determine if alarm view is to be shown
+        // set true to show alarm view
+        if (false) {
+            let storyboard = UIStoryboard(name: "Main", bundle: nil)
+            let vc = storyboard.instantiateViewController(withIdentifier: "AlarmViewController") as! AlarmViewController
+            
+            // TODO: read alarm from DB
+            vc.alarm = Alarm()
+            
+            window?.rootViewController = vc
+        }
+        
         return true
     }
